feat: Update C API to support double precision cluster distances
BREAKING CHANGE: AdaptiveRouteResult.cluster_distance changed from float to double

- C API now properly supports double precision for cluster distances
- Float32 routers return exact values (float→double is lossless)
- Float64 routers preserve full precision
- Required for templated RouteResponseT to work with C bindings
- Existing C code will need to use double instead of float for cluster_distance

+++ adaptive_router_core/bindings/c/include/adaptive.h
@@ -31,11 +31,11 @@ typedef struct AdaptiveRouter AdaptiveRouter;
  * Route result structure
  */
 typedef struct {
-  char* selected_model;      /**< Selected model ID (caller must free with adaptive_string_free) */
-  char** alternatives;       /**< Array of alternative model IDs */
-  size_t alternatives_count; /**< Number of alternatives */
-  int cluster_id;            /**< Assigned cluster ID */
-  float cluster_distance;    /**< Distance to cluster centroid */
+   char* selected_model;      /**< Selected model ID (caller must free with adaptive_string_free) */
+   char** alternatives;       /**< Array of alternative model IDs */
+   size_t alternatives_count; /**< Number of alternatives */
+   int cluster_id;            /**< Assigned cluster ID */
+   double cluster_distance;   /**< Distance to cluster centroid */
 } AdaptiveRouteResult;
 
 /**
